+++ fvm/src/lib.rs
@@ -23,11 +23,17 @@ pub mod state_tree;
 
 mod blockstore;
 
-// TODO Public only for integration tests.
-//  Consider exporting only behind a feature
-pub mod account_actor;
-pub mod init_actor;
-pub mod system_actor;
+cfg_if::cfg_if! {
+    if #[cfg(feature = "testing")] {
+        pub mod account_actor;
+        pub mod init_actor;
+        pub mod system_actor;
+    } else {
+        mod account_actor;
+        mod init_actor;
+        mod system_actor;
+    }
+}
 
 mod market_actor;
 mod power_actor;