+++ shared/CHANGELOG.md
@@ -1,11 +1,18 @@
 # Changelog
 
+## Unreleased
+
+- Added `testing` feature to have `Default` derive on `Message`. Extended this feature to `Address` and `Payload`. 
+
 ## 0.6.0 [2022-04-14]
 
 BREAKING: Switch syscall struct alignment: https://github.com/filecoin-project/fvm-specs/issues/63
 
 Actors built against this new version of fvm_shared will be incompatible with prior FVM versions,
 and vice-versa.
+
+- Added `Display` trait to `Type` for error printing. 
+- Added _cfg = "testing"_ on `Default` trait for `Message` structure.
 
 ## 0.5.1  [2022-04-11]