+++ shared/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
-## Unreleased
+## [Unreleased]
 
-- Added `testing` feature to have `Default` derive on `Message`. Extended this feature to `Address` and `Payload`. 
+## 0.6.1 [2022-04-29]
+
+- Added `testing` feature to have `Default` derive on `Message`. Extended this feature to `Address` and `Payload`.
+- Improve `ErrorNumber` documentation.
+- Update `fvm_ipld_encoding` for the cbor encoder switch.
 
 ## 0.6.0 [2022-04-14]